Rename api to client_api
